refactor(filter): move FilterTriggerButtonRemovable to FilterButton
