+++ packages/components/src/Filter/FilterButton/index.ts
@@ -1,1 +1,2 @@
 export * from "./FilterButton"
+export * from "./FilterButtonRemovable"
