+++ packages/components/src/Filter/components/index.ts
@@ -1,2 +1,2 @@
 export * from "./FilterButton"
-export * from "./FilterButtonRemovable"
+export * from "../FilterButton/FilterButtonRemovable"
